backend/requirements.txt
fastapi
uvicorn
langchain_groq
langchain_community
langchain_core
chromadb
pandas
python-dotenv
beautifulsoup4
